+++ DesktopApplication/DesktopApplication/getremotepacketerrorrateresultscommand.h
@@ -7,15 +7,15 @@ class GetRemotePacketErrorRateResultsCommand : public SerialCommand {
     Q_OBJECT
   public:
     ~GetRemotePacketErrorRateResultsCommand() override;
-    static std::unique_ptr<SerialCommand> create(uint16_t srcAddr, uint16_t dstAddr, uint16_t intervalMs, uint16_t timeout);
-    static SerialCommand* create(QObject *parent, uint16_t srcAddr, uint16_t dstAddr, uint16_t intervalMs, uint16_t timeout);
+    static std::unique_ptr<SerialCommand> create(uint16_t srcAddr, uint16_t dstAddr, uint16_t intervalMs, uint32_t timeout);
+    static SerialCommand* create(QObject *parent, uint16_t srcAddr, uint16_t dstAddr, uint16_t intervalMs, uint32_t timeout);
 
     Stm32SupportedOperations::Stm32SupportedOperationsEnums identifier() const override;
     void iterate(const QDateTime &timestamp, const QString &data) override;
 
   protected:
     explicit GetRemotePacketErrorRateResultsCommand(QObject *parent = nullptr);
-    void initialize(uint16_t srcAddr, uint16_t dstAddr, uint16_t intervalMs, uint16_t timeout) override;
+    void initialize(uint16_t srcAddr, uint16_t dstAddr, uint16_t intervalMs, uint32_t timeout) override;
 };
 
 #endif // GETREMOTEPACKETERRORRATERESULTSCOMMAND_H
